Fast API/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-perplexity


